+++ lib/features/calendar/presentation/views/shared/shared_calendar_screen.dart
@@ -2,7 +2,6 @@ import 'package:dutytable/core/configs/app_colors.dart';
 import 'package:dutytable/core/widgets/back_actions_app_bar.dart';
 import 'package:dutytable/features/calendar/data/models/calendar_model.dart';
 import 'package:dutytable/features/calendar/presentation/viewmodels/shared_calendar_view_model.dart';
-import 'package:dutytable/features/calendar/presentation/views/personal/personal_calendar_screen.dart';
 import 'package:dutytable/features/calendar/presentation/widgets/calender_tab/calendar_tab.dart';
 import 'package:dutytable/features/calendar/presentation/widgets/chat_tab.dart';
 import 'package:dutytable/features/calendar/presentation/widgets/custom_calendar_tabview.dart';
@@ -11,6 +10,8 @@ import 'package:dutytable/features/calendar/presentation/widgets/member_invite_d
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
+
+import '../../widgets/custom_appbar_icon.dart';
 
 class SharedCalendarScreen extends StatelessWidget {
   final CalendarModel calendar;
